feat: add pandas dependency and update test for numpy array formatting

+++ tests/utils/test_formatter.py
@@ -5,7 +5,7 @@ from color_correction.utils.formater import format_value
 
 @pytest.mark.parametrize("value,expected_check", [
     # For numpy array image: expecting an <img src="..."/> tag.
-    (np.array([[1, 2, 3]]), lambda out: out.startswith('<img src="')),
+    (np.array([[1, 2, 3]], dtype=np.uint8), lambda out: out.startswith('<img src="')),
     # For dict: expecting a valid json string that decodes back to the same dict.
     ({"a": 1, "b": 2}, lambda out: json.loads(out) == {"a": 1, "b": 2}),
     # For list: expecting a valid json string that decodes back to the same list.
